Port to async Rust and rustls
OpenSSL has caused nothing but headaches, so let's replace it with
rustls. We also convert our public API to use async.

This patch disables the `stats` and `create_image` features for now,
because we don't use them in `cage`. We commented out the relevant
code instead of deleting it, because we may still want to fix it at
some point.